Deduplicate fake diarize test fixture
